新增用户列表页面

+++ src/components/Home.vue
@@ -1,5 +1,6 @@
 <template>
     <el-container class="home_container">
+        <!-- 头部区域 -->
         <el-header>
             <div class="title">
                 <img src="../assets/logo.png" alt="头像">
@@ -8,21 +9,30 @@
             <el-button type="info" @click="logout" round>退出</el-button>
         </el-header>
         <el-container>
-            <el-aside width="200px">
+            <!-- 侧边栏 -->
+            <el-aside :width="isCollapse?'64px':'200px'">
+                <!-- 缩放侧边栏按钮 -->
+                <div class="collapseBtn" @click="collapseMenu">
+                    <i :class="isCollapse?'el-icon-caret-right':'el-icon-caret-left'"></i>
+                </div>
                 <el-col :span="12" style="width: 100%">
                     <el-menu
                             class="sideMenu"
                             background-color="#f8f9fa"
                             text-color="#333"
-                            active-text-color="#007bff">
+                            active-text-color="#007bff"
+                            unique-opened
+                            router
+                            :collapse="isCollapse"
+                            :collapse-transition="false">
                         <!-- 一级菜单 -->
-                        <el-submenu :index="item.id+''" v-for="item in menuList" :key="item.id">
+                        <el-submenu :index="item.path" v-for="item in menuList" :key="item.id">
                             <template slot="title">
                                 <i class="el-icon-circle-plus"></i>
                                 <span>{{item['authName']}}</span>
                             </template>
                             <!-- 二级菜单 -->
-                            <el-menu-item :index="subItem.id+''" v-for="subItem in item.children" :key="subItem.id">
+                            <el-menu-item :index="subItem.path" v-for="subItem in item.children" :key="subItem.id">
                                 <template slot="title">
                                     <i class="el-icon-ship"></i>
                                     <span>{{subItem['authName']}}</span>
@@ -32,7 +42,10 @@
                     </el-menu>
                 </el-col>
             </el-aside>
-            <el-main>Main</el-main>
+            <!-- 主要内容 -->
+            <el-main>
+                <router-view></router-view>
+            </el-main>
         </el-container>
     </el-container>
 </template>
@@ -42,7 +55,8 @@
         name: "Home",
         data() {
             return {
-                menuList: []
+                menuList: [],
+                isCollapse: false
             }
         },
         created() {
@@ -52,16 +66,26 @@
         methods: {
             //退出登录原理：清空sessionStorage中的token，并手动跳转到 /login 界面
             logout() {
-                window.sessionStorage.clear();
-                this.$router.push('/login');
+                this.$confirm('确定要退出系统吗？', '提示', {
+                    confirmButtonText: '确定',
+                    cancelButtonText: '取消',
+                    type: 'info'
+                }).then(() => {
+                    window.sessionStorage.clear();
+                    this.$router.push('/login');
+                }).catch(() => {
+
+                });
             },
             //获取左侧菜单列表
             async getMenuList() {
                 const {data: res} = await this.$http.get('menus');
                 if (res['meta']['status'] !== 200) return this.$message.error(res['meta']['msg']);
                 this.menuList = res['data'];
-
-                console.log(JSON.stringify(this.menuList));
+            },
+            //缩放侧边栏菜单
+            collapseMenu() {
+                this.isCollapse = !this.isCollapse;
             }
         }
     }
@@ -97,6 +121,10 @@
 
     .el-aside {
         background-color: #f8f9fa;
+
+        .el-menu {
+            border-right: 0;
+        }
     }
 
     .el-main {
@@ -107,4 +135,12 @@
         font-weight: 600;
     }
 
+    .collapseBtn {
+        text-align: center;
+        background-color: #c6c6c6;
+        font-size: 12px;
+        line-height: 17px;
+        letter-spacing: 0.1em;
+        cursor: pointer;
+    }
 </style>
